+++ apps/docs/src/examples/extras/positional-audio/Scene.svelte
@@ -26,10 +26,7 @@
   })
 </script>
 
-<Environment
-  path="/hdr/"
-  files="shanghai_riverside_1k.hdr"
-/>
+<Environment url="/textures/equirectangular/hdr/shanghai_riverside_1k.hdr" />
 
 <T.OrthographicCamera
   {zoom}