+++ apps/docs/src/examples/extras/use-gltf-animations/Tokyo.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
-  import { GLTF, useGltfAnimations } from '@threlte/extras'
+  import { GLTF, useGltfAnimations, useDraco } from '@threlte/extras'
 
+  const dracoLoader = useDraco()
   const { gltf, actions } = useGltfAnimations<'Take 001'>()
 
   $: $actions['Take 001']?.play()
@@ -9,5 +10,5 @@
 <GLTF
   bind:gltf={$gltf}
   url="/models/LittlestTokyo.glb"
-  useDraco="https://www.gstatic.com/draco/v1/decoders/"
+  {dracoLoader}
 />