+++ apps/docs/src/examples/extras/use-gltf/Scene.svelte
@@ -19,10 +19,7 @@
   }>('/models/helmet/DamagedHelmet.gltf')
 </script>
 
-<Environment
-  path="/hdr/"
-  files="shanghai_riverside_1k.hdr"
-/>
+<Environment url="/textures/equirectangular/hdr/shanghai_riverside_1k.hdr" />
 
 <T.PerspectiveCamera
   makeDefault