+++ packages/core/src/lib/index.ts
@@ -1,5 +1,3 @@
-import './types.d.ts'
-
 export const VERSION = 8
 
 // canvas component
@@ -25,6 +23,7 @@ export type { Key, Schedule, Scheduler, Stage, Task, TaskCallback } from './fram
 // useLoader
 export {
   useLoader,
+  type UseLoaderLoadOptions,
   type UseLoaderLoadInput,
   type UseLoaderLoadResult,
   type UseLoaderOptions