+++ packages/core/src/lib/types.d.ts
@@ -1,7 +1,14 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-declare namespace Threlte {
-  interface UserProps {
-    // User can extend with custom props here
+declare global {
+  namespace Threlte {
+    interface UserProps {
+      // User can extend with custom props here
+    }
+    interface UserCatalogue {
+      // User can the <T> catalogue here
+    }
   }
 }
+
+export * from './index'