+++ packages/extras/src/lib/components/HTML/HTML.svelte
@@ -56,6 +56,7 @@
   import type { HTMLProps } from './types'
 
   let {
+    autoRender = true,
     eps = 0.001,
     center = false,
     fullscreen = false,
@@ -124,7 +125,7 @@
     }
   })
 
-  useTask(() => {
+  export const render = () => {
     camera.current.updateMatrixWorld()
     group.updateWorldMatrix(true, false)
     const vec = transform ? oldPosition : calculatePosition(group, camera.current, $size)
@@ -253,6 +254,18 @@
         occlusionMesh.lookAt(camera.current.position)
       }
     }
+  }
+
+  export const { start: startRendering, stop: stopRendering } = useTask(render, {
+    autoStart: false
+  })
+  $effect(() => {
+    if (autoRender) {
+      startRendering()
+      return () => {
+        stopRendering()
+      }
+    }
   })
 
   let pos = $derived.by(() => {
@@ -341,7 +354,7 @@
           class={props.class}
           style={props.style}
         >
-          {@render children?.()}
+          {@render children?.({ render, startRendering, stopRendering })}
         </div>
       </div>
     </div>
@@ -356,7 +369,7 @@
       style={props.style}
       class={props.class}
     >
-      {@render children?.()}
+      {@render children?.({ render, startRendering, stopRendering })}
     </div>
   {/if}
 </svelte:element>