+++ packages/extras/src/lib/components/environment/Environment/Environment.svelte
@@ -1,0 +1,99 @@
+<script
+  lang="ts"
+  module
+>
+  const loaders: {
+    exr?: EXRLoader
+    hdr?: RGBELoader
+    tex?: TextureLoader
+  } = {}
+</script>
+
+<script lang="ts">
+  import { T, useCache, useThrelte } from '@threlte/core'
+  import { EquirectangularReflectionMapping, TextureLoader } from 'three'
+  import { EXRLoader, GroundedSkybox, RGBELoader } from 'three/examples/jsm/Addons.js'
+  import { useSuspense } from '../../../suspense/useSuspense'
+  import { useEnvironment } from '../utils/useEnvironment.svelte'
+  import type { EquirectangularEnvironmentProps } from './types'
+
+  const ctx = useThrelte()
+
+  let {
+    skybox = $bindable(),
+    texture = $bindable(),
+    ground = false,
+    isBackground = false,
+    scene = ctx.scene,
+    url
+  }: EquirectangularEnvironmentProps = $props()
+
+  const suspend = useSuspense()
+  const cache = useCache()
+
+  useEnvironment({
+    get scene() {
+      return scene
+    },
+    get isBackground() {
+      return isBackground
+    },
+    get texture() {
+      return texture
+    }
+  })
+
+  const isEXR = $derived(url?.endsWith('exr') ?? false)
+  const isHDR = $derived(url?.endsWith('hdr') ?? false)
+
+  // defaults to `TextureLoader` if `url` is not provided
+  const loader = $derived.by(() => {
+    if (url === undefined) return
+    if (isEXR) {
+      loaders.exr ??= new EXRLoader()
+      return loaders.exr
+    } else if (isHDR) {
+      loaders.hdr ??= new RGBELoader()
+      return loaders.hdr
+    }
+    loaders.tex ??= new TextureLoader()
+    return loaders.tex
+  })
+
+  $effect(() => {
+    if (url !== undefined && loader !== undefined) {
+      const suspendedTexture = suspend(
+        cache.remember(() => {
+          return loader.loadAsync(url)
+        }, [url])
+      )
+
+      suspendedTexture.then((t) => {
+        t.mapping = EquirectangularReflectionMapping
+        texture = t
+      })
+
+      return () => {
+        suspendedTexture.then((texture) => {
+          texture.dispose()
+        })
+      }
+    }
+  })
+</script>
+
+{#if ground}
+  {@const options = ground === true ? {} : ground}
+  {#if texture}
+    <T
+      is={GroundedSkybox}
+      oncreate={() => {
+        return () => {
+          skybox = undefined
+        }
+      }}
+      bind:ref={skybox}
+      args={[texture, options.height ?? 1, options.radius ?? 1, options.resolution ?? 128]}
+    />
+  {/if}
+{/if}